+++ src/containers/detail/DetailContentsContainer.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react';
 
 import DetailContentsSection from '@/components/detail/DetailContentsSection';
-import useFetchGroup from '@/hooks/api/useFetchGroup';
+import useFetchGroup from '@/hooks/api/group/useFetchGroup';
 
 function DetailContentsContainer(): ReactElement {
   const { data: group } = useFetchGroup();
