+++ src/hooks/api/group/useFetchUserAppliedGroups.ts
