+++ src/hooks/api/group/useFetchUserRecruitedGroups.test.ts
@@ -1,30 +1,30 @@
 import { renderHook } from '@testing-library/react-hooks';
 
-import { getUserAppliedGroups } from '@/services/api/applicants';
+import { getUserRecruitedGroups } from '@/services/api/group';
 import wrapper from '@/test/ReactQueryWrapper';
 
-import FIXTURE_GROUP from '../../../fixtures/group';
+import FIXTURE_GROUP from '../../../../fixtures/group';
 
-import useFetchUserAppliedGroups from './useFetchUserAppliedGroups';
+import useFetchUserRecruitedGroups from './useFetchUserRecruitedGroups';
 
-jest.mock('@/services/api/applicants');
+jest.mock('@/services/api/group');
 
-describe('useFetchUserAppliedGroups', () => {
-  const useFetchUserAppliedGroupsHook = () => renderHook(() => useFetchUserAppliedGroups('userUid'), {
+describe('useFetchUserRecruitedGroups', () => {
+  const useFetchUserRecruitedGroupsHook = () => renderHook(() => useFetchUserRecruitedGroups('userUid'), {
     wrapper,
   });
 
   beforeEach(() => {
     jest.clearAllMocks();
 
-    (getUserAppliedGroups as jest.Mock).mockImplementation(() => (given.groups));
+    (getUserRecruitedGroups as jest.Mock).mockImplementation(() => (given.groups));
   });
 
   context('useQuery 반환값이 존재하지 않는 경우', () => {
     given('groups', () => null);
 
     it('빈 배열을 반환해야만 한다', async () => {
-      const { result, waitFor } = useFetchUserAppliedGroupsHook();
+      const { result, waitFor } = useFetchUserRecruitedGroupsHook();
 
       await waitFor(() => !!result.current.data);
 
@@ -36,7 +36,7 @@ describe('useFetchUserAppliedGroups', () => {
     given('groups', () => [FIXTURE_GROUP]);
 
     it('groups에 대한 정보를 반환해야만 한다', async () => {
-      const { result, waitFor } = useFetchUserAppliedGroupsHook();
+      const { result, waitFor } = useFetchUserRecruitedGroupsHook();
 
       await waitFor(() => !!result.current.data);
