+++ src/pages/index.test.tsx
@@ -1,8 +1,8 @@
 import { render } from '@testing-library/react';
 import { useRouter } from 'next/router';
 
-import useFetchGroups from '@/hooks/api/useFetchGroups';
-import useFetchTagsCount from '@/hooks/api/useFetchTagsCount';
+import useFetchGroups from '@/hooks/api/group/useFetchGroups';
+import useFetchTagsCount from '@/hooks/api/tagsCount/useFetchTagsCount';
 import InjectTestingRecoilState from '@/test/InjectTestingRecoilState';
 
 import GROUP_FIXTURE from '../../fixtures/group';
@@ -12,8 +12,8 @@ import HomePage from './index.page';
 jest.mock('next/router', () => ({
   useRouter: jest.fn(),
 }));
-jest.mock('@/hooks/api/useFetchGroups');
-jest.mock('@/hooks/api/useFetchTagsCount');
+jest.mock('@/hooks/api/group/useFetchGroups');
+jest.mock('@/hooks/api/tagsCount/useFetchTagsCount');
 
 describe('HomePage', () => {
   beforeEach(() => {
